+++ bolt11/index.js
@@ -1,8 +1,12 @@
+const byteDecodeRequest = require('./byte_decode_request');
+const byteEncodeRequest = require('./byte_encode_request');
 const createSignedRequest = require('./create_signed_request');
 const createUnsignedRequest = require('./create_unsigned_request');
 const parsePaymentRequest = require('./parse_payment_request');
 
 module.exports = {
+  byteDecodeRequest,
+  byteEncodeRequest,
   createSignedRequest,
   createUnsignedRequest,
   parsePaymentRequest,