+++ index.js
@@ -1,8 +1,12 @@
+const {byteDecodeRequest} = require('./bolt11');
+const {byteEncodeRequest} = require('./bolt11');
 const {createSignedRequest} = require('./bolt11');
 const {createUnsignedRequest} = require('./bolt11');
 const {parsePaymentRequest} = require('./bolt11');
 
 module.exports = {
+  byteDecodeRequest,
+  byteEncodeRequest,
   createSignedRequest,
   createUnsignedRequest,
   parsePaymentRequest,